requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
DateTime==4.2
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
zope.interface==4.5.0
